Implementados os logs quando renomeado, Trocado o Softlink por Hardlink ao renomear

+++ MediaManager/Commands/RenomearCommand.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Runtime.InteropServices;
 using System.Windows;
 using System.Windows.Input;
 using MediaManager.Helpers;
@@ -32,36 +33,52 @@ namespace MediaManager.Commands
             {
                 if (item.IsSelected)
                 {
+                    Helper.LogMessage("O arquivo \"" + Path.Combine(item.FolderPath, item.Filename) + "\" será copiado para \"" + Path.Combine(item.Serie.FolderPath, item.FilenameRenamed) + "\"");
+                    Helper.LogMessage("Método de processamento: " + ((Enums.MetodoDeProcessamento)Properties.Settings.Default.pref_MetodoDeProcessamento).ToString());
+
                     if (!Directory.Exists(item.Serie.FolderPath))
+                    {
                         Directory.CreateDirectory(item.Serie.FolderPath);
+                        Helper.LogMessage("Diretório \"" + item.Serie.FolderPath + "\" criado.");
+                    }
+
                     if (File.Exists(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed)))
                     {
-                        if (MessageBox.Show("O episódio " + item.FilenameRenamed + " já existe. Você deseja sobrescrevê-lo pelo arquivo \"" + Path.Combine(item.FolderPath, item.Filename) + "\"?", Properties.Settings.Default.AppName, MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
+                        if (!renomearVM.IsSilencioso)
                         {
-                            File.Delete(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed));
-                            if (Helper.CreateSymbolicLink(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed),
-                                    Path.Combine(item.FolderPath, item.Filename), Enums.SymbolicLink.File))
+                            if (MessageBox.Show("O episódio " + item.FilenameRenamed + " já existe. Você deseja sobrescrevê-lo pelo arquivo \"" + Path.Combine(item.FolderPath, item.Filename) + "\"?", Properties.Settings.Default.AppName, MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
                             {
-                                item.FilePath = Path.Combine(item.Serie.FolderPath, item.FilenameRenamed);
-                                item.IsRenamed = true;
-                                DBHelper.UpdateEpisodioRenomeado(item);
+                                File.Delete(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed));
+                                if (Helper.CreateHardLink(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed),
+                                        Path.Combine(item.FolderPath, item.Filename), IntPtr.Zero))
+                                {
+                                    item.FilePath = Path.Combine(item.Serie.FolderPath, item.FilenameRenamed);
+                                    item.IsRenamed = true;
+                                    DBHelper.UpdateEpisodioRenomeado(item);
+                                }
+                                else
+                                {
+                                    Helper.TratarException(new Exception("Código: " + Marshal.GetLastWin32Error() + "\r\nArquivo: " + Path.Combine(item.FolderPath, item.Filename)), "Ocorreu um erro ao criar o HardLink.\r\nDetalhes: ", true);
+                                }
                             }
-                            else // TODO Excluir
-                            {
-                            }
+                        }
+                        else
+                        {
+                            Helper.LogMessage("O arquivo " + Path.Combine(item.Serie.FolderPath, item.FilenameRenamed) + " já existe e não será sobrescrito.");
                         }
                     }
                     else
                     {
-                        if (Helper.CreateSymbolicLink(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed),
-                                Path.Combine(item.FolderPath, item.Filename), Enums.SymbolicLink.File))
+                        if (Helper.CreateHardLink(Path.Combine(item.Serie.FolderPath, item.FilenameRenamed),
+                                Path.Combine(item.FolderPath, item.Filename), IntPtr.Zero))
                         {
                             item.FilePath = Path.Combine(item.Serie.FolderPath, item.FilenameRenamed);
                             item.IsRenamed = true;
                             DBHelper.UpdateEpisodioRenomeado(item);
                         }
-                        else // TODO Excluir
+                        else
                         {
+                            Helper.TratarException(new Exception("Código: " + Marshal.GetLastWin32Error() + "\r\nArquivo: " + Path.Combine(item.FolderPath, item.Filename)), "Ocorreu um erro ao criar o HardLink.\r\nDetalhes: ", true);
                         }
                     }
                 }
